requirements.txt
# ----------------------------------------------------
# Web 框架与服务器 (Web Framework & Server)
# ----------------------------------------------------
fastapi==0.115.12
uvicorn==0.34.3
starlette==0.46.2
sse-starlette==3.0.3
bottle==0.13.4
python-multipart==0.0.21

# ----------------------------------------------------
# 网络请求与异步 IO (Network Requests & Async IO)
# ----------------------------------------------------
aiohttp==3.13.1
httpx==0.28.1
requests==2.32.4
websockets==15.0.1
anyio==4.9.0
greenlet==3.2.2
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx-sse==0.4.3
idna==3.10
multidict==6.7.0
propcache==0.4.1
pycares==4.11.0
sniffio==1.3.1
urllib3==2.4.0
yarl==1.22.0
certifi==2025.4.26
charset-normalizer==3.4.2
dnspython==2.8.0

# ----------------------------------------------------
# 数据库与缓存 (Database & Cache)
# ----------------------------------------------------
SQLAlchemy==2.0.41
alembic==1.15.2
mysql-connector-python==9.3.0
redis==7.1.0
fakeredis==2.33.0
diskcache==5.6.3
py-key-value-aio==0.3.0
py-key-value-shared==0.3.0

# ----------------------------------------------------
# AI 与大模型 (AI & LLM)
# ----------------------------------------------------
openai==2.6.0
mcp==1.21.0
fastmcp==2.8.1
jiter==0.11.1
openapi-pydantic==0.5.1
proxy_tools==0.1.0

# ----------------------------------------------------
# 数据解析与处理 (Data Parsing & Processing)
# ----------------------------------------------------
beautifulsoup4==4.13.4
lxml==5.4.0
soupsieve==2.7
numpy==2.3.4
PyYAML==6.0.3
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
docutils==0.22.4
docstring_parser==0.17.0
markdown-it-py==4.0.0
mdurl==0.1.2
referencing==0.36.2
rpds-py==0.30.0
pandas==2.3.3
openpyxl==3.1.5

# ----------------------------------------------------
# 模板引擎 (Templating)
# ----------------------------------------------------
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2

# ----------------------------------------------------
# 监控与观测 (Observability & Monitoring)
# ----------------------------------------------------
loguru==0.7.3
python-json-logger==4.0.0
opentelemetry-api==1.39.1
opentelemetry-exporter-prometheus==0.60b1
opentelemetry-instrumentation==0.60b1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
prometheus_client==0.23.1

# ----------------------------------------------------
# 安全与加密 (Security & Cryptography)
# ----------------------------------------------------
cryptography==46.0.3
pycryptodome==3.23.0
Authlib==1.6.6
PyJWT==2.10.1
keyring==25.7.0
email-validator==2.3.0
cffi==2.0.0
pycparser==2.23

# ----------------------------------------------------
# 系统、桌面应用与打包 (System, Desktop & Packaging)
# ----------------------------------------------------
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
pywebview==5.1
pyperclip==1.11.0
altgraph==0.17.5
macholib==1.16.4
platformdirs==4.5.1
shellingham==1.5.4
pathable==0.4.4
pathvalidate==3.3.1
cyclopts==4.4.3
pydocket==0.16.3
distro==1.9.0
lupa==2.6

# macOS specific pyobjc
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-Quartz==12.1
pyobjc-framework-Security==12.1
pyobjc-framework-UniformTypeIdentifiers==12.1
pyobjc-framework-WebKit==12.1

# ----------------------------------------------------
# 通用工具与核心依赖 (Utilities & Core Dependencies)
# ----------------------------------------------------
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
typer==0.21.0
click==8.2.0
rich==14.2.0
rich-rst==1.3.2
Pygments==2.19.2
python-dotenv==1.1.0
schedule==1.2.2
tqdm==4.67.1
cachetools==5.5.2
attrs==25.4.0
typing_extensions>=4.15.0
annotated-types==0.7.0
typing-inspection==0.4.0
beartype==0.22.9
cloudpickle==3.1.2
exceptiongroup==1.3.1
importlib_metadata==8.7.1
zipp==3.23.0
more-itertools==10.8.0
packaging==25.0
six==1.17.0
sortedcontainers==2.4.0
wrapt==1.17.3
jaraco.classes==3.4.0
jaraco.context==6.0.2
jaraco.functools==4.4.0
setuptools==80.9.0
colorama==0.4.6

# ----------------------------------------------------
# 时间与日期 (Date & Time)
# ----------------------------------------------------
pytz==2025.2
python-dateutil==2.9.0.post0
tzdata==2025.2

# ----------------------------------------------------
# 第三方服务 SDK (Third Party Services)
# ----------------------------------------------------
alibabacloud-oss-v2==1.1.1
ccxt==4.5.12
crcmod-plus==2.1.0

# ----------------------------------------------------
# 浏览器模拟 (Browser Simulation)  PyExecJS 解决滑块验证问题
# ----------------------------------------------------
playwright==1.57.0
PyExecJS==1.5.1

# ----------------------------------------------------
# 异步文件操作 (Async File Operations)
# ----------------------------------------------------
aiofiles==25.1.0